Fix Windows ARM64 native builds and runtime portability
- Teach Windows demo builds to target the host architecture by default instead of hardcoding x64, while preserving explicit arm64/x64 overrides through --arch and VIPER_DEMO_ARCH.
- Extend Windows ARM64 native linking with MSVC runtime/debug helper implementations, ExitProcess startup import planning, ARM64 COFF relocation constants, section-relative relocation application, and dynamic import handling needed by full demo builds.
- Harden cross-platform runtime behavior across atomics, threading, networking, TLS verification, crypto helpers, localization, number formatting, file/archive IO, graphics, D3D11, VM bridging, and bytecode paths.
- Update build configuration and Windows build scripts so ARM64 toolchain/runtime selection works without regressing x64, Linux, or macOS paths.
- Refresh demo and smoke coverage for 3dbowling, 3dscene, and baseball persistence after the native build fixes.
- Add and update tests covering linker imports/archive behavior, runtime networking/TLS/formatting, tool install smoke checks, and related CMake wiring.
- Include generated Visual Studio/CMake x64 Debug metadata under out/build/x64-Debug as requested so all tracked and untracked files are captured.

Verification performed before commit:
- cmake --build build --config Debug --target viper -j 8 -- /v:minimal
- scripts\build_demos.cmd --clean
- Confirmed all generated demo executables report PE machine 0xAA64 on Windows ARM64.
- ctest --test-dir build -C Debug -R "linker|runtime_import|archive" --output-on-failure

+++ src/codegen/common/linker/CoffReader.cpp
@@ -359,7 +359,8 @@ bool readCoffObj(
                 return false;
             }
             rel.symIndex = cr->SymbolTableIndex + 1; // +1 because ObjFile has null sym at 0.
-            rel.addend = extractCoffAddend(hdr->Machine, rel.type, sec.data, rel.offset);
+            rel.addend =
+                extractCoffAddend(hdr->Machine, static_cast<uint16_t>(rel.type), sec.data, rel.offset);
 
             sec.relocs.push_back(rel);
         }
